Restructure the navbar
