+++ src/App.js
@@ -1,7 +1,7 @@
 import { Route, Routes } from 'react-router-dom';
 import './App.css';
 import Navbar from './components/NavBar/Navbar';
-import Rockets from './components/Rockets';
+import Rockets from './components/Rocket/Rockets';
 import MyProfile from './components/ProfilePage/MyProfile';
 import Missions from './components/Mission/Missions';
